server/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==9.0.0
psycopg2-binary==2.9.3
